requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
tqdm>=4.66.2
pandas>=2.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0